+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentTable/PersistentDataPanel.tsx
@@ -19,14 +19,12 @@ import PersistentToolBar from "../PersistentToolBar/PersistentToolBar";
 import PersistentTableData from "./PersistentTableData";
 import PersistentDataGridHeaders from "./PersistentDataGridHeaders";
 import PersistentVSCodeAPI from "../PersistentVSCodeAPI";
-// import * as l10n from "@vscode/l10n";
 
 export default function PersistentDataPanel({ type }: Readonly<{ type: Readonly<string> }>): JSXInternal.Element {
   const [data, setData] = useState<{ [type: string]: { [property: string]: string[] } }>({ ds: {}, uss: {}, jobs: {} });
   const [selection, setSelection] = useState<{ [type: string]: string }>({ [type]: "search" });
   const [persistentProp, setPersistentProp] = useState<string[]>([]);
   const [selectedItems, setSelectedItems] = useState({});
-  // const [localization, setLocalization] = useState(null);
 
   const selectedItemsMemo = useMemo(
     () => ({
@@ -57,13 +55,6 @@ export default function PersistentDataPanel({ type }: Readonly<{ type: Readonly<
       if (!isSecureOrigin(event.origin)) {
         return;
       }
-      // if (event.data.command === "GET_LOCALIZATION") {
-      //   const { contents } = event.data;
-      //   l10n.config({
-      //     contents: contents,
-      //   });
-      //   setLocalization(contents);
-      // }
       setData(event.data);
 
       if ("selection" in event.data) {
@@ -72,7 +63,6 @@ export default function PersistentDataPanel({ type }: Readonly<{ type: Readonly<
         }));
       }
     });
-    // PersistentVSCodeAPI.getVSCodeAPI().postMessage({ command: "GET_LOCALIZATION" });
   }, []);
 
   useEffect(() => {